requirements.txt
streamlit
pandas
plotly
pillow
scikit-learn
torch
torchvision